requirements.txt
pyperclip
requests
Pillow